Refactor and new ease of use feature
